requirements-core.txt
# Основные зависимости для работы парсера (HTTP-only режим)
# Для LIGHT режима - достаточно только этих зависимостей

# HTTP клиенты
aiohttp>=3.9.0
curl-cffi>=0.5.0

# Логирование
loguru>=0.7.0

# Обработка данных
pandas>=2.0.0
openpyxl>=3.1.0

# Конфигурация
python-dotenv>=1.0.0

# Windows специфичные (только для Windows)
pywin32>=306; sys_platform == "win32"
